requirements.txt
streamlit
transformers
torch
faiss-cpu
sentence_transformers
numpy
requests
pandas